updateLayout
